+++ changelog.md
@@ -1,4 +1,4 @@
 [![Downloads](https://img.shields.io/github/downloads/zevnda/steam-game-idler/1.6.6/total?style=for-the-badge&logo=github&color=137eb5)](https://github.com/zevnda/steam-game-idler/releases/download/1.6.6/Steam.Game.Idler_1.6.6_x64_en-US.msi)
 
 ## Changelog
-- Fixed intervals for `card farming`
+- Fixed an issue where `achievement unlocker` wasn't correctly checking for private games or profile